feat: add options
무기 회전 속도, 데미지, 칼 회전 간격 조정 선택지

+++ Great Swordsman/Assets/Scripts/PlayerHP.cs
@@ -21,7 +21,7 @@ public class PlayerHP : MonoBehaviour
         UpdateHealthBar();
     }
 
-    public void IncreaseHP(string type, float value)
+    public void IncreaseHP(string type, int value)
     {
         if (type == "currentHP")
         {
